PoP precision: more bugfixes
There were issues with how the debug detection was getting the variable back
from the stored value. Also, a scoping problem on startTime caused an issue
when debug was enabled.

Text handling for understanding how close a chosen team is to the 7h minimum
was only working previously by luck, it is more robust now, as is the
condition it needs to match.